websocket client factory
