+++ packages/adapters/eth/Cargo.toml
@@ -36,6 +36,7 @@ thiserror = { workspace = true }
 tokio = { workspace = true, features = ["time"] }
 tracing = { workspace = true }
 url = { workspace = true }
+trait-variant = { workspace = true }
 
 [dev-dependencies]
 tracing-subscriber = { workspace = true, features = [
