+++ packages/adapters/eth/src/lib.rs
@@ -8,14 +8,14 @@ mod metrics;
 mod provider;
 mod websocket;
 
-pub use std::sync::Arc;
 pub use alloy::primitives::Address;
 pub use aws::*;
 pub use blob_encoder::BlobEncoder;
 pub use error::{Error, Result};
 pub use failover_client::FailoverClient;
 pub use http::Provider as HttpClient;
-pub use provider::{L1Provider, ProviderConfig, ProviderFactory, create_real_provider_factory};
+pub use provider::L1Provider;
+pub use std::sync::Arc;
 pub use websocket::{
     AcceptablePriorityFeePercentages, L1Key, L1Keys, Signer, Signers, TxConfig, WebsocketClient,
 };
