+++ packages/adapters/eth/src/provider.rs
@@ -5,18 +5,19 @@ use futures::future::BoxFuture;
 use std::ops::RangeInclusive;
 use std::sync::Arc;
 use std::time::Duration;
+use url::Url;
 
 use crate::error::{Error as EthError, Result as EthResult};
+use crate::failover_client::{ProviderConfig, ProviderInit};
 use crate::{L1Keys, TxConfig, WebsocketClient};
 use services::state_committer::port::l1::Priority;
 use services::types::{
     BlockSubmissionTx, Fragment, FragmentsSubmitted, L1Tx, NonEmpty, TransactionResponse, U256,
 };
 
-/// Trait defining the necessary L1 provider operations
-/// This is implemented by WebsocketClient and mocks for testing
+#[allow(async_fn_in_trait)]
+#[trait_variant::make(Send)]
 #[cfg_attr(test, mockall::automock)]
-#[async_trait]
 pub trait L1Provider: Send + Sync {
     /// Get the current block number from the L1 network
     async fn get_block_number(&self) -> EthResult<u64>;
@@ -52,30 +53,7 @@ pub trait L1Provider: Send + Sync {
     async fn submit(&self, hash: [u8; 32], height: u32) -> EthResult<BlockSubmissionTx>;
 }
 
-/// A struct that holds configuration for a provider
-#[derive(Clone, Debug)]
-pub struct ProviderConfig {
-    /// Optional name for logging purposes
-    pub name: String,
-    /// The URL to connect to
-    pub url: String,
-}
-
-impl ProviderConfig {
-    pub fn new(name: impl Into<String>, url: impl Into<String>) -> Self {
-        Self {
-            name: name.into(),
-            url: url.into(),
-        }
-    }
-}
-
-/// A factory that can create an L1Provider from a ProviderConfig
-pub type ProviderFactory<P> =
-    Box<dyn Fn(&ProviderConfig) -> BoxFuture<'static, EthResult<Arc<P>>> + Send + Sync + 'static>;
-
 /// Implementation of L1Provider for WebsocketClient
-#[async_trait]
 impl L1Provider for WebsocketClient {
     async fn get_block_number(&self) -> EthResult<u64> {
         Ok(self._get_block_number().await?)
@@ -120,7 +98,6 @@ impl L1Provider for WebsocketClient {
 }
 
 /// Blanket implementation for references to types that implement L1Provider
-#[async_trait]
 impl<T: L1Provider + ?Sized> L1Provider for &T {
     async fn get_block_number(&self) -> EthResult<u64> {
         (*self).get_block_number().await
@@ -165,38 +142,37 @@ impl<T: L1Provider + ?Sized> L1Provider for &T {
     }
 }
 
-/// Creates a factory that produces real WebsocketClient instances
-pub fn create_real_provider_factory(
+pub struct WebsocketClientFactory {
     contract_address: Address,
     signers: Arc<crate::websocket::Signers>,
     unhealthy_after_n_errors: usize,
     tx_config: TxConfig,
-) -> ProviderFactory<WebsocketClient> {
-    Box::new(move |config: &ProviderConfig| {
-        let contract_address = contract_address;
-        let signers = signers.clone();
-        let tx_config = tx_config.clone();
+}
+
+impl ProviderInit for WebsocketClientFactory {
+    type Provider = WebsocketClient;
+
+    async fn initialize(&self, config: &ProviderConfig) -> EthResult<Arc<Self::Provider>> {
+        let contract_address = self.contract_address;
+        let signers = self.signers.clone();
+        let tx_config = self.tx_config.clone();
         let url_str = config.url.clone();
 
-        Box::pin(async move {
-            use url::Url;
+        // Parse the URL
+        let url =
+            Url::parse(&url_str).map_err(|e| EthError::Other(format!("Invalid URL: {}", e)))?;
 
-            // Parse the URL
-            let url =
-                Url::parse(&url_str).map_err(|e| EthError::Other(format!("Invalid URL: {}", e)))?;
+        // Create the WebsocketClient
+        let client = WebsocketClient::connect(
+            url,
+            contract_address,
+            (*signers).clone(),
+            self.unhealthy_after_n_errors,
+            tx_config,
+        )
+        .await
+        .map_err(|e| EthError::Other(format!("Failed to connect: {}", e)))?;
 
-            // Create the WebsocketClient
-            let client = WebsocketClient::connect(
-                url,
-                contract_address,
-                (*signers).clone(),
-                unhealthy_after_n_errors,
-                tx_config,
-            )
-            .await
-            .map_err(|e| EthError::Other(format!("Failed to connect: {}", e)))?;
-
-            Ok(Arc::new(client))
-        })
-    })
+        Ok(Arc::new(client))
+    }
 }
